feat: Remove UserHomePage and add user-related components
- Deleted UserHomePage component.
- Added AppointmentsComponent to display upcoming appointments.
- Introduced EmergencyComponent for searching blood donors.
- Created HistoryComponent to show blood donation history.
- Added InvitationsComponent for managing blood donation invitations.
- Implemented ProfileComponent for user profile management.
- Developed RemindersComponent for blood donation reminders.
